requirements-original.txt
streamlit==1.32.0
python-docx==1.1.0
PyPDF2==3.0.1
langchain==0.1.12
chromadb==0.4.24
openai==1.14.3
httpx<0.27
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
tiktoken==0.6.0
pinecone-client==2.2.4
fpdf==1.7.2